Fix segfault compiler bug.
Before this fix, a seg fault would always occur with dpdk, 4 cores, 512 outstanding packets, and a tensor with 255 elements or 64255 elements. Disabling optimization fixes that bug which may suggest that it is indeed a compiler bug.
Regardless, while the exact reason remains unknown, adding this line seems to fix the issue for now.