Exponents support in P4 program, SDE 9.6 support
Moved to SDE 9.6.
The P4 program now supports the processing of exponents with a max
operation.
With RDMA the immediate value contains a message ID (16 bits) and
2 exponents (8 bits each). The message ID is sent back untouched
(must be the same in corresponding packets from all the workers).
With UDP the 2 exponents (8 bits each) are immediately after the
SwitchML header.